+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the Solarium library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Ureleased]
+### Fixed
+- Facet pivot stats contain all fields (previously only the last field was present in the result)
+- Facet pivot stats return `NAN` for a mean value that's NaN
+- Facet pivot stats return an associative array for percentiles
+
+### Changed
+- Facet pivot stats results use the field name instead of `'stats_fields'` as array key
+
+
 ## [6.3.3]
 ### Changed
 - Spatial component distance type changed from int to float